Changes to JavaEncrytor CTOR, and hash(), sign(), verifySignature() member functions to account for preferred byte encoding (defaults to UTF-8) rather using than native platform byte encoding that can vary from one platform to the next.
Updated main()'s javadoc and changed System.out.println from using "\n" to
the Java property line.separator which should pick the correct line
terminator as per native OS. Defaults to newline in unusual event that
line.separator property is not set.

Completed Javadoc for main().

NB: These changes could conceivably affect those that were running on a
platform where UTF-8 is not the native encoding (e.g., Windows OS) resulting
in a different master secret. If this is a problem we can revert back to previous
version or user could change encoding in ESAPI.properties. However, I'd rather
cause issues now rather than later when many more are using it.  I still need to
check the EncryptorTest. I think it will be OK with these changes, but have not
yet verified that it didn't write something out to a file that it later reads back
in. If so, that may break too.