SafeFileTest assumed that ${user.home} ended with a directory separator and failed if it did not.
Instead of fixing this directly, I changed the test to use a temporary
file using File#createTempFile() and added a finally to properly remove
it after testing. This also fixed the test failing.

This whole test class needs this cleanup in a similar manner as creating
test files outside the build directory, except temporary files, is a
rather bad idea. If temporary files are used, they should be properly
cleaned up as well. Created issue 56 detailing this as I'm trying to
fix unit test right now.