Fixes for Encryptedproperties, DefaultEncryptedProperties and EncryptedPropertiesTest in 1.4. These will be migrated to 2.0 in my next commit.
Modify DefaultEncrypedProperties#getProperty(String) to return null when
the key does not exist. This is more inline with what users will expect
as it is what java.util.Properties#getProperty(String) does. Previously
this would throw a NullPointerException in Base64#decode(String) when
it tried to decode null which was confusing at best.

Modify javadoc for EncryptedProperties#getProperty(String) to define
the expected behavior in the case of a non-existent key.

Add EncryptedPropertiesTest#testNonExistantKey() to test the behavior
of non-existent keys in isolation.

Modify EncryptedPropertiesTest#testGetProperty() to not expect an
Exception to be thrown in the case of a non-existant key.

Modify EncrypedPropertiesTest#testKeySet() to not depend on the order
of the keys in the key set.

Combine EncrypedPropertiesTest#testStore()
and EncryptedPropertiesTest#testLoad() into
EncryptedProperties#testStoreLoad() as testLoad() depended on testStore()
running first which I'm not sure junit/surefire guarantees. Also modify
to write to and read from a byte array input and output stream to avoid
managing a temporary file.

Remove EncryptedProperties#main(String[]) as it wasn't worth porting the
above to it and mvn -Dtest=EncryptedPropertiesTest test is functionally
equivalent to what was originally desired.

I think that's all...