Delete org.eclipse.jdt.core.prefs
This is a file generated by Eclipse and not really part of ESAPI per se. It may be version specific to a particular release of Eclipse, thus we are deleting it and adding it to be ignored in the .gitignore file. (In fact, this one is telling Eclipse to use '-target 1.5' when actually '-target 1.6' is what the pom.xml specifies.)