Whitelist Regex Validation Tests
Providing a structure to assert that the enviroment configurations will
provide expected responses when passed controlled values.

Performing a preliminary whitelist test set on HttpQueryString regex from
ESAPI.properties.

The test implementation asserts that the regex being tested matches the
test environment's regex for a given property. If the regex strings
differ, the tests associated with that property will be skipped to prevent
false-positives from being provided to a client.