Removing unstable tests
The tests removed from SafeRequestTest were attempting to validate
environment state as well as the unit functionality.  The intent of these
tests has been distributed to a workflow test check
(SecurityWrapperRequestTest), a whitelist validation test
(EsapiWhitelistValidationPatternTester), and tests for the PercentCodec's
impact on input data (PercentCodecStringTest, PercentCodecCharacterTest).

The combination of these tests assert that the pieces that were composing
the removed content will function under more explict and controlled
conditions.  Changes to configuration will nChanges to configuration will
now be revealed closer to the component that is directly impacted.