POM UPDATES (dependency refs & versions)
Increasing versions of dependencies.

The updated dependency for commons-collections removed ArrayListIterator
and CollectionsUtil from the library.  Those classes appear to have been
moved to org.apache.commons:commons-collections4.  I swapped the
dependency out to allow for the version update.

Similar case with the powermock-api-mockito.  A class was removed from the
referenced library.  Replaced reference with the latest
powermock-api-mockito2.

WAS NOT ABLE TO UPDATE
javax.servlet:javax.servlet-api from 3.0.1 to 4.0.1
Modification adds API to ServletInputStream and ServletOutputStream, and
requires work in the ESAPI 'waf' packages to compensate.