Log special (#451)
* Close issue #448

* Close issue #444. Delete deprecated decodeToObject() method and 2 related encodeObject() methods. General whitespace clean-up. Delete EncoderTest.testBase64decodToObject() method which is no longer relevant.

* Close issue #385.  Close issue #386.
NOTE: Was unwilling to comply with request in these 2 issues to make logSpecial()
'protected'. Could not do so without introducing potential security vulnerabilities.
However, since the intent of the user submitting these 2 GitHub issues was only to override
logSpecial() in order to completely suppress the output from them to stdout or stderr,
I have arranged it so that setting the System property
'org.owasp.esapi.logSpecial.discard' to 'true' will do exactly this...it will
suppress all output from logSpecial. (Also, the calls to System.err.println() and
System.out.println() have been replaced by calls to logSpecial(), which will log
to System.out if not suppressed. So the end result is all or nothing. I suspect
that most will keep the default behavior, which is to print to System.out rather
than suppressing all output.