Authenticatortest Stability
In addition to issues resolved in previous commit, an additional
intermittent problem was identified having to do with file IO during
testing.  To account for the IO timing, the timeout of the implementation
was extended from 10 seconds to 5 minutes.  The timeout was originally
applied to prevent the test from running indefinetly, without
consideration for any performance requirements.  The timeout extension
still fulfills the original intent and accounts for issues we're currently
experiencing.