DateValidationRule Logic Updates
Updating the parsing logic of the DateValidationRule to not only verify a
date can be constructed from the input string in a known format, but also
that the generated Date can be reformatted back to the same String
content.  This adds additional security for cases where the DateFormatter
accepts characters postfixed to an otherwise valid date.

As a result two tests in ValidationTest.java had to be updated to use the
correct explicit format.  The data sets were using full-name month, day,
year content, but supplying the MEDIUM format. Updating the provided
format to LONG allowed the new check to function as desired.