AbstractAccessReferenceMap Issues #37 #329 (#469)
* Concurrency Test for AbstractAccessReferenceMap

SThis test case consistently illustrates the race condition resulting from
concurrent invocations of addDirectReference.

* AbstractAccessReferenceMap Sync Updates

Synchronizing all public access methods to this wrapper class.

* AbstractAccessReferenceMap Class Documentation

Updating class documentation to provide an example of multi-interaction
synchronization syntax.

* AbstractAccessMap field updates to use HashMap

With primary synchronization being pushed to the class API level, there
really is no benefit to maintaining the overhead of ConcurrentHashMap
fields.  The thread-safety guaranteed by the ConcurrentHashMap is
superceeded by the public API of the abstract class, and should never
actually be leveraged.

This assumes that all sub-classes with extended public API continue the
use of class-level synchronization.

* AbstractAccessReferenceMapTest Extension

Adding method which shows that it is possible to create a corrupted direct
-> Indirect to Indirect -> Direct Object relationship when using the
public update(Set) API.

* AbstractAccessReferenceMap.update Logic Update

Altering the logic of update behavior to avoid the possibility of the same
indirect object being associated with multiple direct instances.