MySQLCodec Logic & Tests invalid int CTR Arg
Found that implementations would throw NPE's if an invalid int was
provided to the deprecated constructor when trying to encode or decode
values.  Opted to alter the constructor logic to immediately throw a
runtime exception (IllegalArgument) if the constructor parameter cannot
resolve to a valid reference.

It would be a Runtime exception either way, it's just a matter of when
it's thrown.

Alternatively, the mode switches in the encode/decode could have been
surrounded by an if block with an else/fallthrough which retuned null.

It is my opinion that this is a non-obvious failure of misconfiguration
and would be difficult to diagnose and resolve.  The immediate class
failure should provide the end-user with context on how to resolve the
problem and limit the time debugging.

Tests provided to validate workflow.