MySQLCodec Code Structure Updates
Making a copy of MySQLCodec in its own subpackage of the codec group.  Breaking
apart the implementation into the codec and a strategy pattern controlled
by the MySQLMode enumeration.

Each piece of this implementation retained the original tests; however,
like the code, the tests have been split up as well.  Now the
responsibilty of the ANSI handing is in the MySQLAnsiSupport class.
Standard behavior is held an tested from the MySQLStandardSupport class.

I feel like this change increases readability and usability of the
baseline.  I am committing this at this time to get another set of
opinions on this approach.  In the interim, the original MySQLCodec and
tests have remained untouched in this effort.