DefaultValidator input Validation tests
Test for getValidInput verifying that input is correctly forwarded to the
StringValidationRule, result from the StringValidationRule is returned on
success, and that any ValidationExceptions during the process are handled
as intended.

I am not writing tests for the use of the canonicalize parameter for these
methods at this time.  Presently there is no associated functionality or
workflow for that element so it's difficult to verify anything more than
the value I passed into the method is present in the scope of the method.