Issue #188 resolution proof:  Test updates (#475)
[NOTE: The main purpose of this PR was to provide enough tests to so that GitHub issue #188  is now operating as desired and can be closed. It is not certain _when_ this issue was fixed, but the intent of these tests is to so that 188 is no longer a problem. - @kwwall ]

* SecurityWrapperRequestTest Extension: getParameter

Adding the set of conditional tests for the getParameter(String)

* SecurityWrapperRequestTest getParameter HappyPaths

Creating happy-path tests for verifying getParameter calls through to
Validator.getValidInput as desired

* SecurityWrapperResponseTest getParameter cont

Creating tests to verify result on null reference being handled and
behavior on ValidationException for the remaining getParameter API.

* SecurityWrapperRequestTest cleanup

Extracting Strings to test-scope members.  Test name is used for
uniqueness between tests.

* SecurityWrapperRequestTest Cleanup

Reducing duplicated code in test blocks by abstracting out interaction
with the Validator interface to a support class.

* SecurityWrapperRequestTest Consistency Cleanup

Applying the delegate validator interaction class to the tests for
QueryString.

* DefaultValidator input Validation tests

Test for getValidInput verifying that input is correctly forwarded to the
StringValidationRule, result from the StringValidationRule is returned on
success, and that any ValidationExceptions during the process are handled
as intended.

I am not writing tests for the use of the canonicalize parameter for these
methods at this time.  Presently there is no associated functionality or
workflow for that element so it's difficult to verify anything more than
the value I passed into the method is present in the scope of the method.

* DefaultValidator isValidInput tests

Extending the valid input testing to include the isValid methods.

* DefaultValidator ValidInput cleanup: Imports

Using static imports for junit Assert and Mockito.