Slf4JLogFactory Test Updates & Fnctl changes
The constructor of the LogPrefixAppender was updated to keep
the interaction with the ESAPI properties consistent between the
Log Appender and the Log Scrubber.  Now all property lookups are held
exclusively in static block of the SLF4JLogFactory and passed into the
implementation.

Slf4JLogFactoryTests updated to account for API updates.