IGNORED ClientInfoSupplierTest
Creating initial stub of ClientInfoSupplier Test.  I simply cannot get the
User interface to mock for the test. Likewise, I cannot create a
DefaultUser.  There's something off with the Thread context. I think it is
because the User interface extends java.security.Principal, but I'm really
not sure.

Tests are incomplete and will need to be updated once this issue is
overcome.

Presently content is ignored with a FIXME in the impl.