JavaLogFactory Failed Configuration Load Test
Updating the JavaLogFactory to have a delegate static method for loading
the configuration.  Allowing the LogManager to be provided as an argument
to avoid the need for verbose mocking and setup in the test scope.

Adding a test that asserts that Standard Error stream is provided the
expected error content if the configuration cannot be resolved.