Intermittent Test Failure Cleanup
The property manager and property loader tests were altering system
properties used to initialize common resources. The original values were
not restored, and I was having test failures from the system being unable
to resolve the values configured for the tests.

I have added static BeforeAll and AfterAll blocks to the three tests to
capture the original property state before the test runs, and to put it
back after all tests are finished.