HTMValidationRule & Test Updates (#607)
* HTMLValidationRule static updates

Splitting the logic of the static block into multiple smaller blocks.
Doing this appears to have make the tests more consistent between the full
build and running the ClasspathTest in isolation.

Both now fail due to bad configuration in the esapi-antisamy-CP.xml file.

* HTML Validation CP Tests

Correcting test file reference to point to the antisamy copy file.

Removing invalid characters from the antisamy copy file so that it may be
parsed by the runtime.

* HTMLValidationRule Tests

Adding a known invalid antisamy configuration file.

Updating the static initialization blocks from private to
package-protected for targeted testing.

Adding a test that asserts that when we attempt to use an invalid Antisamy
configuration file, that a PolicyException is emitted.

* Adding Test for Antisamy Schema Validation

TEST IS FAILING

Adding test to verify that the known bad configuration is accepted when
the system property for Antisamy is set appropriately.

* Updating Antisamy Version

Update to 1.6.1

* Antisamy Stream WorkAround

Creating a ByteArrayInputStream in the HTMLValidationRule to circumvent
the undesired stream close event when processing a malformed xml
configuraiton while antisamy schema validation is disabled.

Moving the property test to an isolated file.

* Antisamy Schema Validation Tests

Switching to configuring validation behavior through static methods rather
than system properties.  Properties are read once on construction in the
Policy, and without forcing test execution order, the validation would be
based on execution order

* Stream to ByteArrayStream Fix

Updates to the process to convert the resource stream to a
ByteArrayInputStream through the use of a Reader implementation.

This allows the java7 try-with-resources block when converting the
InputStream Object, which should ensure that the original stream is
cleaned up once read into memory.

* Cleanup

Removing dead method.
Adding newline at end of file

* Antisamy 1.6.2 Update

Prepping content for the Antisamy 1.6.2 artifact that is expected to be
available in the near future.

Updated pom dependency to anticipated version

Updated HTMLValidationRule to remove pre-wrapping of ByteArrayInputStream
for Policy Object creation

* HTMLValidationRule Cleanup

Removing unused imports.