commons-fileupload v1.4
Updating dependency and correcting failed test in HTTPUtilitiesTest.

Updated version was failing when trying to overwrite an existing file on
the filesystem.
Looked into the FileItem API, and found that the file system reference
is removed when the FileItem is GC'd. Replicated that intent from the
test scope and forcibly deleted files created (and returned) by
DefaultHTTPUtilities.

This may still be a problem for a deployed instance if a user tries to
upload a file multiple times, or multiple users try to upload a file by
the same name within the same GC cycle.