Configuring mvn versions plugin
Enabling an explicit exclusion of the xml-apis dependency when checking
for updates.  As noted in the file, xml-apis rolled backwards in
versioning (starting at 2.x and rolling backwards to 1.x).

This will prevent the version check from showing any xml-apis 2.x
version as newer than our current project configuration, which is the
correct behavior based on release dates in maven.